Refactoring #2. Changed names, removed unnecessary Toast
