+++ app/src/main/java/com/example/playlistmaker/data/local/database/TrackListenHistoryLocalDataSource.kt
@@ -3,34 +3,34 @@ package com.example.playlistmaker.data.local.database
 import com.example.playlistmaker.data.model.Track
 
 class TrackListenHistoryLocalDataSource(
-    private val localDatabase: LocalDatabase
+    private val localDataSource: LocalDataSource
 ) : DataSource {
     override fun loadAllTracks(): List<Track> {
-        val result = localDatabase.loadListenHistoryTracks().toMutableList()
+        val result = localDataSource.loadListenHistoryTracks().toMutableList()
         result.reverse()
         return result
     }
 
     override fun saveAllTracks(listOfTracks: List<Track>) {
-        localDatabase.saveListenHistoryTracks(listOfTracks.toTypedArray())
+        localDataSource.saveListenHistoryTracks(listOfTracks.toTypedArray())
     }
 
     override fun clearDatabase() {
-        val tracks = localDatabase.loadListenHistoryTracks().toMutableList()
+        val tracks = localDataSource.loadListenHistoryTracks().toMutableList()
         tracks.clear()
-        localDatabase.saveListenHistoryTracks(tracks.toTypedArray())
+        localDataSource.saveListenHistoryTracks(tracks.toTypedArray())
     }
 
     override fun removeTrack(id: Long) {
-        val tracks = localDatabase.loadListenHistoryTracks().toMutableList()
+        val tracks = localDataSource.loadListenHistoryTracks().toMutableList()
         for (track in tracks) {
             if (track.trackId == id) tracks.remove(track)
         }
-        localDatabase.saveListenHistoryTracks(tracks.toTypedArray())
+        localDataSource.saveListenHistoryTracks(tracks.toTypedArray())
     }
 
     override fun loadTrack(id: Long): Track? {
-        val tracks = localDatabase.loadListenHistoryTracks().toMutableList()
+        val tracks = localDataSource.loadListenHistoryTracks().toMutableList()
         var result: Track? = null
         for (track in tracks) {
             if (track.trackId == id) result = track
@@ -40,7 +40,7 @@ class TrackListenHistoryLocalDataSource(
 
     override fun saveTrack(track: Track) {
         val tracks: MutableList<Track> =
-            localDatabase.loadListenHistoryTracks().toMutableList()
+            localDataSource.loadListenHistoryTracks().toMutableList()
         var result = mutableListOf<Track>()
         if (tracks.isEmpty()) {
             result.add(track)
@@ -54,11 +54,11 @@ class TrackListenHistoryLocalDataSource(
                 result.add(track)
             }
         }
-        localDatabase.saveListenHistoryTracks(result.toTypedArray())
+        localDataSource.saveListenHistoryTracks(result.toTypedArray())
     }
 
     fun getSize(): Int {
-        return localDatabase.loadListenHistoryTracks().toMutableList().size
+        return localDataSource.loadListenHistoryTracks().toMutableList().size
     }
 
 }
