+++ app/src/main/java/com/example/playlistmaker/data/repositorie/settingRepository/SettingsRepository.kt
@@ -1,8 +1,8 @@
 package com.example.playlistmaker.data.repositorie.settingRepository
 
-import com.example.playlistmaker.data.local.database.LocalDatabase
+import com.example.playlistmaker.data.local.database.LocalDataSource
 
-class SettingsRepository(private val database: LocalDatabase) {
+class SettingsRepository(private val database: LocalDataSource) {
     fun saveBooleanSetting(key: String, value: Boolean) {
         database.saveBoolean(key, value)
     }
